+++ test/integration/adapters/smtp_test.exs
@@ -5,8 +5,8 @@ defmodule Swoosh.Integration.Adapters.SMTPTest do
 
   @moduletag integration: true
 
-  setup_all do
-    config = [
+  defp config(Swoosh.Adapters.SMTP) do
+    [
       relay: System.get_env("SMTP_RELAY"),
       username: System.get_env("SMTP_USERNAME"),
       password: System.get_env("SMTP_PASSWORD"),
@@ -14,40 +14,57 @@ defmodule Swoosh.Integration.Adapters.SMTPTest do
       tls: :always,
       auth: :always
     ]
-
-    {:ok, config: config}
   end
 
-  test "simple deliver", %{config: config} do
-    email =
-      new()
-      |> from({"Swoosh SMTP", "swoosh+smtp@#{config[:domain]}"})
-      |> reply_to("swoosh+replyto@#{config[:domain]}")
-      |> to("swoosh+to@#{config[:domain]}")
-      |> cc("swoosh+cc@#{config[:domain]}")
-      |> bcc("swoosh+bcc@#{config[:domain]}")
-      |> subject("Swoosh - SMTP integration test")
-      |> text_body("This email was sent by the Swoosh library automation testing")
-      |> html_body("<p>This email was sent by the Swoosh library automation testing</p>")
-
-    assert {:ok, _response} = Swoosh.Adapters.SMTP.deliver(email, config)
+  defp config(Swoosh.Adapters.Muaua) do
+    [
+      relay: System.get_env("SMTP_RELAY"),
+      domain: System.get_env("SMTP_DOMAIN"),
+      auth: [
+        username: System.get_env("SMTP_USERNAME"),
+        password: System.get_env("SMTP_PASSWORD")
+      ]
+    ]
   end
 
-  test "deliver with attachment in memory", %{config: config} do
-    email =
-      new()
-      |> from({"Swoosh SMTP", "swoosh+smtp@#{config[:domain]}"})
-      |> to("swoosh+to@#{config[:domain]}")
-      |> subject("Swoosh - SMTP integration test")
-      |> text_body("This email was sent by the Swoosh library automation testing")
-      |> attachment(%Swoosh.Attachment{
-        content_type: "text/plain",
-        data: "this is an attachment",
-        filename: "example.txt",
-        type: :attachment,
-        headers: []
-      })
+  for adapter <- [Swoosh.Adapters.SMTP, Swoosh.Adapters.Mua] do
+    describe "using #{adapter}" do
+      setup do
+        {:ok, config: config(unquote(adapter))}
+      end
 
-    assert {:ok, _response} = Swoosh.Adapters.SMTP.deliver(email, config)
+      test "simple deliver", %{config: config} do
+        email =
+          new()
+          |> from({"Swoosh SMTP", "swoosh+smtp@#{config[:domain]}"})
+          |> reply_to("swoosh+replyto@#{config[:domain]}")
+          |> to("swoosh+to@#{config[:domain]}")
+          |> cc("swoosh+cc@#{config[:domain]}")
+          |> bcc("swoosh+bcc@#{config[:domain]}")
+          |> subject("Swoosh - SMTP integration test")
+          |> text_body("This email was sent by the Swoosh library automation testing")
+          |> html_body("<p>This email was sent by the Swoosh library automation testing</p>")
+
+        assert {:ok, _response} = unquote(adapter).deliver(email, config)
+      end
+
+      test "deliver with attachment in memory", %{config: config} do
+        email =
+          new()
+          |> from({"Swoosh SMTP", "swoosh+smtp@#{config[:domain]}"})
+          |> to("swoosh+to@#{config[:domain]}")
+          |> subject("Swoosh - SMTP integration test")
+          |> text_body("This email was sent by the Swoosh library automation testing")
+          |> attachment(%Swoosh.Attachment{
+            content_type: "text/plain",
+            data: "this is an attachment",
+            filename: "example.txt",
+            type: :attachment,
+            headers: []
+          })
+
+        assert {:ok, _response} = unquote(adapter).deliver(email, config)
+      end
+    end
   end
 end